Make the extension work with the new event system in PHPUnit 10
