+++ .github/workflows/build.yml
@@ -15,10 +15,10 @@ jobs:
         name: Build and test
         strategy:
             matrix:
-                php: [8.0, 8.1, 8.2]
+                php: [8.1, 8.2]
                 deps: [high]
                 include:
-                    -   php: 8.0
+                    -   php: 8.1
                         deps: low
 
         steps:
@@ -45,7 +45,7 @@ jobs:
 
             -   name: Upload the phar
                 uses: actions/upload-artifact@v1
-                if: matrix.php == '8.0' && matrix.deps == 'high'
+                if: matrix.php == '8.1' && matrix.deps == 'high'
                 with:
                     name: zalas-phpunit-globals-extension.phar
                     path: build/zalas-phpunit-globals-extension.phar
@@ -56,8 +56,7 @@ jobs:
         needs: tests
         strategy:
             matrix:
-                php: [8.0, 8.1, 8.2]
-
+                php: [8.1, 8.2]
         steps:
             -   uses: actions/checkout@v3
