+++ composer.json
@@ -10,11 +10,8 @@
         }
     ],
     "require": {
-        "php": "~8.0.0 || ~8.1.0 || ~8.2.0",
-        "phpunit/phpunit": "^9.0"
-    },
-    "conflict": {
-        "phpunit/phpunit": "9.5.0"
+        "php": "~8.1.0 || ~8.2.0",
+        "phpunit/phpunit": "^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -26,9 +23,11 @@
             "Zalas\\PHPUnit\\Globals\\Tests\\": "tests/"
         }
     },
+    "prefer-stable": true,
+    "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "2.x-dev"
+            "dev-master": "3.x-dev"
         }
     }
 }
