+++ manifest.xml.in
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <phar xmlns="https://phar.io/xml/manifest/1.0">
     <contains name="zalas/phpunit-globals" version="@@version@@" type="extension">
-        <extension for="phpunit/phpunit" compatible="^9.0"/>
+        <extension for="phpunit/phpunit" compatible="^10.0"/>
     </contains>
 
     <copyright>
@@ -10,6 +10,6 @@
     </copyright>
 
     <requires>
-        <php version="^8.0"/>
+        <php version="^8.1"/>
     </requires>
 </phar>
