+++ phpunit.xml.dist
@@ -1,10 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd" bootstrap="vendor/autoload.php" beStrictAboutOutputDuringTests="true" beStrictAboutTodoAnnotatedTests="true" colors="true" verbose="true">
-  <coverage processUncoveredFiles="true">
-    <include>
-      <directory suffix=".php">src</directory>
-    </include>
-  </coverage>
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="file://./vendor/phpunit/phpunit/phpunit.xsd" bootstrap="vendor/autoload.php" beStrictAboutOutputDuringTests="true" beStrictAboutTodoAnnotatedTests="true" colors="true">
   <testsuite name="default">
     <directory suffix="Test.php">tests</directory>
     <exclude>tests/phar</exclude>
@@ -13,6 +8,6 @@
     <env name="USER" value="test" force="true"/>
   </php>
   <extensions>
-    <extension class="Zalas\PHPUnit\Globals\AnnotationExtension"/>
+    <bootstrap class="Zalas\PHPUnit\Globals\AnnotationExtension" />
   </extensions>
 </phpunit>
