+++ src/GlobalsBackup.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+namespace Zalas\PHPUnit\Globals;
+
+use PHPUnit\Event\Test\Prepared;
+use PHPUnit\Event\Test\PreparedSubscriber;
+
+final class GlobalsBackup implements PreparedSubscriber
+{
+    private $server;
+    private $env;
+    private $getenv;
+
+    public function notify(Prepared $event): void
+    {
+        $this->server = $_SERVER;
+        $this->env = $_ENV;
+        $this->getenv = \getenv();
+    }
+
+    public function restoreGlobals(): void
+    {
+        $_SERVER = $this->server;
+        $_ENV = $this->env;
+
+        foreach (\array_diff_assoc($this->getenv, \getenv()) as $name => $value) {
+            \putenv(\sprintf('%s=%s', $name, $value));
+        }
+        foreach (\array_diff_assoc(\getenv(), $this->getenv) as $name => $value) {
+            \putenv($name);
+        }
+    }
+}
